requirements.txt
matplotlib==3.0.3
numpy==1.19.5
pandas==1.1.5
Pillow==8.2.0
scikit-learn==0.24.1
scipy==1.5.4
torch==1.0.1.post2
visualdl==2.1.1

# for reading the tf model
tensorflow-gpu==1.15.0
